Silence compiler warning

+++ generic/tkBitField.c
@@ -127,7 +127,7 @@ LsbIndex(uint32_t x)
 	 0,  1, 28,  2, 29, 14, 24, 3, 30, 22, 20, 15, 25, 17,  4, 8,
 	31, 27, 13, 23, 21, 19, 16, 7, 26, 12, 18,  6, 11,  5, 10, 9
     };
-    return MultiplyDeBruijnBitPosition[((uint32_t) ((x & -x)*0x077cb531)) >> 27];
+    return MultiplyDeBruijnBitPosition[((uint32_t) ((x & -((int32_t) x))*0x077cb531)) >> 27];
 }
 #  else
 /* The "classical" method, but about 20% slower than the DeBruijn method on my system. */
